first draft of the quiz
The quiz first draft could be reshulfed to a smooth logical order, and can be reviewd based on what they mostly need to remember for the project. It can have some more conceptual questions. 

But it is review most of the topics covered in the firsta day and fucntionalities they need to remeber!